Part 6: Acceptance testing. Tasks

+++ app/controllers/questions_controller.rb
@@ -2,10 +2,11 @@ class QuestionsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
   
   expose :questions, ->{ Question.all }
-  expose :question
+  expose :question  
+  expose :answer, ->{ Answer.new }
 
   def create
-        
+    @exposed_question = current_user.questions.new(question_params)
     if question.save
       redirect_to question, notice: 'Your question successfully created.'
     else      
@@ -22,7 +23,10 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user.author_of?(question)
+      question.destroy
+      flash[:notice] = 'The question is successfully deleted.'
+    end
     redirect_to questions_path
   end
